feat(using): 添加对自定义类dispose方法的支持
- 在AnyLangValue中实现TryDispose方法，支持自定义类的资源释放
- 修改UsingStatement以支持AnyLangValue类型的资源清理
- 更新ValueTypeFuncStatic中的类型转换逻辑，修复默认值问题
- 优化DictionaryValueFuncStatic中的VariateManager获取方式
- 修复TupleValueFuncStatic中的元素遍历逻辑，避免递归栈溢出
- 添加test_disposable.old8测试文件验证using语句的dispose功能

+++ Old8Lang/AST/Expression/ValueFunctions/CharValueFuncStatic.cs
@@ -111,7 +111,7 @@ public static class CharValueFuncStatic
         /// <returns>字符的整数值</returns>
         public IntLangValue ToInt()
         {
-            return IntLangValue.Create((int)charValue.Value);
+            return IntLangValue.Create(charValue.Value);
         }
 
         /// <summary>
